+++ test/dummy/app/models/order.rb
@@ -2,9 +2,9 @@ class Order < ApplicationRecord
   belongs_to :customer
 
   janela do
-    measure :revenue, sum: :amount
+    measure :revenue, sum: :amount, prefix: "$"
     measure :orders, count: true
-    measure :average_order, average: :amount
+    measure :average_order, average: :amount, prefix: "$"
 
     dimension :status
     dimension :region, through: :customer